requirements.txt
Flask
aiosqlite
discord.py
PyNaCl
python-dotenv
motor
